requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
sqlalchemy
psycopg2-binary
celery
redis
requests
asyncpg
google-generativeai
